Added UnitControls.cs
Moved MonoBehaviour class outside of UnitSpawner.cs script
