+++ Age Of War - 4D Edition/Assets/Scripts/UnitSpawner.cs
@@ -71,19 +71,3 @@ public class UnitSpawner : MonoBehaviour
         }
     }
 }
-
-public class UnitControls : MonoBehaviour
-{
-    public Unit unitData;
-
-    void Update()   // temporary code to test movement
-    {
-        gameObject.transform.position += Vector3.right * unitData.movementSpeed * Time.deltaTime;
-
-        if(gameObject.transform.position.x >= 25)
-        {
-            Destroy(gameObject);
-        }
-    }
-}
-
